backend/requirements.txt
# FastAPI and web server
fastapi==0.120.2
uvicorn[standard]==0.32.1
python-multipart==0.0.18

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Authentication and security
pyjwt==2.10.1
passlib==1.7.4
bcrypt==4.0.1

# OpenAI API (Primary - Production RAG)
openai==1.57.2

# Qdrant vector database client
qdrant-client==1.15.1

# Document processing
PyMuPDF==1.26.5
python-docx==1.1.2

# Data validation
pydantic==2.10.3
pydantic[email]==2.10.3

# Environment variables
python-dotenv==1.0.1

# Utilities
numpy==1.26.4

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0
